:beer:  [plaoc]  修复app之间的消息通信

+++ next/kmp/browser/src/commonMain/kotlin/org/dweb_browser/browser/jsProcess/ext/createJsProcess.kt
@@ -97,6 +97,6 @@ class JsProcess(
         parameters["globalId"] = remoteGlobalId.toString()
         parameters["manifest"] = Json.encodeToString(remoteMM)
       }.buildUnsafeString()
-    ).int()
+    )
   }
 }
